Subplot int arguments

+++ osimpipeline/postprocessing.py
@@ -307,7 +307,7 @@ def plot_marker_error_general(output_filepath, marker_names, ymax, gl,
 
     fig = pl.figure(figsize=(12, 4 * np.ceil(len(marker_names) * 0.5)))
     for imark, marker_name in enumerate(marker_names):
-        pl.subplot(np.ceil(len(marker_names) * 0.5), 2, imark + 1)
+        pl.subplot(int(np.ceil(len(marker_names) * 0.5)), 2, int(imark + 1))
         if (marker_name[0] == '.' or marker_name[0] == '_' 
             or marker_name[0]=='*'):
